Update per comments
