+++ kyuubi-ha/src/main/scala/org/apache/kyuubi/ha/client/DiscoveryClientProvider.scala
@@ -21,6 +21,8 @@ import java.io.IOException
 
 import org.apache.kyuubi.{KyuubiException, Logging}
 import org.apache.kyuubi.config.KyuubiConf
+import org.apache.kyuubi.ha.HighAvailabilityConf
+import org.apache.kyuubi.util.ClassUtils
 
 object DiscoveryClientProvider extends Logging {
 
@@ -43,15 +45,15 @@ object DiscoveryClientProvider extends Logging {
 
   def createDiscoveryClient(conf: KyuubiConf): DiscoveryClient = {
     val classLoader = Thread.currentThread.getContextClassLoader
-    val className = conf.get(KyuubiConf.DISCOVERY_CLIENT_CLASS)
+    val className = conf.get(HighAvailabilityConf.DISCOVERY_CLIENT_CLASS)
     if (className.isEmpty) {
       throw new KyuubiException(
-        s"${KyuubiConf.DISCOVERY_CLIENT_CLASS.key} cannot be empty.")
+        s"${HighAvailabilityConf.DISCOVERY_CLIENT_CLASS.key} cannot be empty.")
     }
     val cls = Class.forName(className, true, classLoader)
     cls match {
       case c if classOf[DiscoveryClient].isAssignableFrom(cls) =>
-        KyuubiConf.createInstance[DiscoveryClient](c, conf)
+        ClassUtils.createInstance[DiscoveryClient](c, conf)
       case _ => throw new KyuubiException(
         s"$className must extend of ${DiscoveryClient.getClass.getName}")
     }
